[CIDMDEV-5131] - Migration Support - Read-Only Client Role
- Added flag feature.migration.read.only.mode.enabled to enable read-only mode
- Protected the following services by flag (returns 503 when read-only enabled) :
POST v2.0/OS-KSADM/roles
DELETE v2.0/OS-KSADM/roles/{roleId}

+++ src/main/java/com/rackspace/idm/api/resource/cloud/v20/Cloud20VersionResource.java
@@ -6,7 +6,10 @@ import com.rackspace.docs.identity.api.ext.rax_kskey.v1.ApiKeyCredentials;
 import com.rackspace.docs.identity.api.ext.rax_ksqa.v1.SecretQA;
 import com.rackspace.idm.JSONConstants;
 import com.rackspace.idm.api.serviceprofile.CloudContractDescriptionBuilder;
+import com.rackspace.idm.domain.config.IdentityConfig;
 import com.rackspace.idm.exception.BadRequestException;
+import com.rackspace.idm.exception.ExceptionHandler;
+import com.rackspace.idm.exception.MigrationReadOnlyIdmException;
 import com.rackspace.idm.exception.NotFoundException;
 import com.rackspace.idm.util.SamlUnmarshaller;
 import org.apache.commons.configuration.Configuration;
@@ -40,6 +43,9 @@ public class Cloud20VersionResource {
     private Configuration config;
 
     @Autowired
+    private IdentityConfig identityConfig;
+
+    @Autowired
     private CloudContractDescriptionBuilder cloudContractDescriptionBuilder;
 
     private static final String X_AUTH_TOKEN = "X-AUTH-TOKEN";
@@ -58,6 +64,9 @@ public class Cloud20VersionResource {
 
     @Autowired
     private CloudMultifactorDomainResource multifactorDomainResource;
+
+    @Autowired
+    private ExceptionHandler exceptionHandler;
 
     public Cloud20VersionResource() {
     }
@@ -674,7 +683,11 @@ public class Cloud20VersionResource {
             @Context UriInfo uriInfo,
             @HeaderParam(X_AUTH_TOKEN) String authToken,
             Role role) {
-        return cloud20Service.addRole(httpHeaders, uriInfo, authToken, role).build();
+        if (identityConfig.getReloadableConfig().migrationReadOnlyEnabled()) {
+            return exceptionHandler.exceptionResponse(new MigrationReadOnlyIdmException()).build();
+        } else {
+            return cloud20Service.addRole(httpHeaders, uriInfo, authToken, role).build();
+        }
     }
 
     @GET
@@ -692,7 +705,11 @@ public class Cloud20VersionResource {
             @Context HttpHeaders httpHeaders,
             @HeaderParam(X_AUTH_TOKEN) String authToken,
             @PathParam("roleId") String roleId) {
-        return cloud20Service.deleteRole(httpHeaders, authToken, roleId).build();
+        if (identityConfig.getReloadableConfig().migrationReadOnlyEnabled()) {
+            return exceptionHandler.exceptionResponse(new MigrationReadOnlyIdmException()).build();
+        } else {
+            return cloud20Service.deleteRole(httpHeaders, authToken, roleId).build();
+        }
     }
 
     @GET
